# blueprints/industry/healthcare-triage/specification/api-contracts.yaml
openapi: 3.0.3
info:
  title: Healthcare AI Triage System API
  description: |
    Comprehensive API for intelligent patient triage and care coordination.
    
    This API provides endpoints for:
    - Patient registration and management
    - AI-powered triage assessment
    - Multi-agent interaction and coordination
    - Clinical decision support
    - Real-time communication and alerts
    - HIPAA-compliant audit and reporting
    
    **Security**: All endpoints require authentication and follow HIPAA compliance standards.
    **Performance**: 95th percentile response time < 500ms for all endpoints.
    **Availability**: 99.9% uptime SLA with automated failover.
  version: 1.0.0
  contact:
    name: Healthcare Triage API Support
    email: api-support@healthcare-triage.com
    url: https://docs.healthcare-triage.com
  license:
    name: Proprietary
    url: https://healthcare-triage.com/license
servers:
  - url: https://api.healthcare-triage.com/v1
    description: Production server
  - url: https://staging-api.healthcare-triage.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
security:
  - BearerAuth: []
  - HIPAACompliant: []
paths:
...
  /knowledge/search:
    post:
      summary: Search medical knowledge base
      description: Perform semantic search across medical protocols, guidelines, and drug interactions
      operationId: searchKnowledge
      security:
        - BearerAuth: []
        - ClinicalRole: []
      tags:
        - Knowledge Base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeSearch'
            examples:
              protocol_search:
                summary: Search for treatment protocols
                value:
                  query: "chest pain treatment protocol emergency department"
                  filters:
                    content_type: ["protocols", "guidelines"]
                    specialty: ["emergency_medicine", "cardiology"]
                    urgency_level: ["high", "critical"]
                  limit: 10
      responses:
        '200':
          description: Knowledge search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeSearchResults'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
    HIPAACompliant:
      type: apiKey
      in: header
      name: X-HIPAA-Audit-Token
      description: HIPAA compliance tracking token
    ClinicalRole:
      type: oauth2
      description: OAuth2 scope for clinical staff access
      flows:
        authorizationCode:
          authorizationUrl: https://auth.healthcare-triage.com/oauth/authorize
          tokenUrl: https://auth.healthcare-triage.com/oauth/token
          scopes:
            clinical:read: Read access to clinical data
            clinical:write: Write access to clinical data
            patient:read: Read access to patient information
            patient:write: Write access to patient information
    AdminRole:
      type: oauth2
      description: OAuth2 scope for administrative access
      flows:
        authorizationCode:
          authorizationUrl: https://auth.healthcare-triage.com/oauth/authorize
          tokenUrl: https://auth.healthcare-triage.com/oauth/token
          scopes:
            admin:read: Read access to administrative functions
            admin:write: Write access to administrative functions
            system:monitor: System monitoring access
    ManagerRole:
      type: oauth2
      description: OAuth2 scope for management access
      flows:
        authorizationCode:
          authorizationUrl: https://auth.healthcare-triage.com/oauth/authorize
          tokenUrl: https://auth.healthcare-triage.com/oauth/token
...
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        correlation_id:
          type: string
          format: uuid
          description: Unique identifier for tracking this error
...
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation_error:
              summary: Input validation error
              value:
                error: "validation_error"
                message: "Invalid input parameters"
                details:
                  field: "medical_record_number"
                  issue: "Must match pattern MRN-YYYY-NNNNNN"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_token:
              summary: Missing authentication token
              value:
                error: "authentication_required"
                message: "Valid authentication token required"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            insufficient_permissions:
              summary: Insufficient permissions
              value:
                error: "insufficient_permissions"
                message: "Clinical role required for this operation"
...